accountadapter: clear filter when adding a contact
When you search for a new contact and clic on "add to conversations"
it will trigger a ContactAdded signal. However, now LRC is async and
for swarm the conversation can take some ms to generate. This means
that the conversation will not exists after ContactAdded for swarm.
Just reset when a new contact is added.

Change-Id: I4ca28be7f2b63b75ba66b7d28268721e714b0e78