calloverlays: cleanup main and participant fadeout mechanisms
Unfortunately, QQuickMouseArea force accepts all move events by
design, which makes passthrough events tricky to catch. In-call
overlays for conference participants should be a layer below the
main overlay. So, using QML MouseAreas makes correctly implementing
both overlay fade mechanisms with passthrough for participant menus
too complex. To properly implement the signal relay mechanism we
would need to listen to all the child components with MouseAreas
and relay the positionChanged events to parenting MouseAreas.

To avoid the complexity, we can install an event filter for the
main overlay and listen to global mouse move events that are
contained in it's geometry.

Additionally, we can use Control components instead of transparent
Rectangles to access a hovered property which can be used to
prevent the fade out, and effectively freeze the visibility of menu
components.

Gitlab: #420
Gitlab: #421
Gitlab: #423
Change-Id: I7f05d2bc79ad6a8db0951b7b59e52f671fdf4798