mainapplication: add "Show Jami" tray icon context menu item
On some desktop environments on GNU/Linux, left-clicking on the tray
icon does not execute the restore action.  In some environments, such
as Ubuntu and some other GNOME-based environments, double-clicking the
icon does that, but that's counter-intuitive and not obvious at all to
the user.  So, we add a 'Show Jami' menu item for the system tray
context menu, similar to how jami-gnome did, so the user can clearly
see it as a possible action.

Change-Id: I1e12ba02190c12cf40d293d61e1f023380fbfcf9