smartlist: filter conversation update/reselection
Filter either the same conv id or peer list of the currently
selected item when responding to the conversationReady signal.

For dialogs we make sure the peer list(1) is same and for all other
swarm conversations we match the conv id.

This will prevent programmatic auto-selection of the smartlist
during account sync.

Depends on: https://review.jami.net/c/ring-lrc/+/18793

Change-Id: I6cedb8494f7c001bf886fe36dd03adffd201bb51