smartlist: improve cache cleaning and avoid glitches
Interaction's cache is cleaned to avoid to store too much informations
and to refresh preferences for the chatview when necessary.
However, this was called too much times. We only need to clear cache
when the settings are changing or if we need to remove a lot of
conversations. This patch uses the SmartListItem to determine
what conversation needs to be cleaned.

Change-Id: I731bb9e7d41b140bbc2607800c1a8e0a1eff3244