mainwindow: don't save fullscreen geometry
This covers the case where closing the app while in fullscreen mode
will save the window geometry. A patch was already made to prevent
the client from restoring fullscreen visibility, however, restoring
the geometry will stretch the window over the screen giving the
impression that it is fullscreen. This patch fixes that behavior.

Change-Id: I520d528a0d8fb62c84bfd79d2f2229bcc654bf8f