deleteaccountdialog: disable delete button when it's clicked
This is so that the user cannot accidentally click Delete again and
delete another account.  Also, show a BusyIndicator spinning wheel
when Delete is clicked, so as to give the user visual feedback that
something is happening.

Also, set the account Id, best name, and whether it's a SIP account
from outside when opening the dialog, so that they don't automatically
change when the next account is selected if the user confirms the
deletion of the current account.

GitLab: #711
Change-Id: I4e93deb4f74257387884edc62d76a5b3cb58d845