misc: vendor libjamiclient (i.e. LRC) into 'src/lib/'
This is libjamiclient from the jami-libclient.git repository as of
c5bca2933fb02ef7f81e8fe0772801a4f97fd1f5, with the following changes:

* src/lib/newaccountmodel.cpp (NewAccountModelPimpl::removeFromAccounts):
The lock wait inside `#ifdef CHK_FREEABLE_BEFORE_ERASE_ACCOUNT' was
not updated when the type of `accounts' was changed over the years
from this:
  std::map<std::string, account::Info>
to this:
  std::map<QString, std::pair<account::Info, std::shared_ptr<Database>>>
Basically we need to get the `first' of the pair for `account::Info'.
So we now do that.

* src/lib/avmodel.cpp:
* src/lib/callbackshandler.cpp:
* src/lib/contactmodel.cpp:
* src/lib/conversationmodel.cpp:
* src/lib/database.cpp:
* src/lib/namedirectory.cpp:
* src/lib/newaccountmodel.cpp:
* src/lib/newcallmodel.cpp:
* src/lib/newdevicemodel.cpp:
* src/lib/peerdiscoverymodel.cpp:
* src/lib/pluginmodel.cpp:
* src/lib/smartinfohub.cpp:
* src/lib/vcard.h:
* src/lib/authority/storagehelper.cpp:  Replace Qt's `foreach' with
`Q_FOREACH' and its `emit' with `Q_EMIT', because in the client-qt
code base we have `-DQT_NO_KEYWORDS' to avoid conflicts with other
libraries we use.

GitLab: #734
Change-Id: I86e3b0fb30e554755023e7b858b6a0d132cd59ab