misc: vendor libjamiclient into 'src/libclient/'
This is libjamiclient from the jami-libclient.git repository as of
c5bca2933fb02ef7f81e8fe0772801a4f97fd1f5, with the following changes:

* src/libclient/newaccountmodel.cpp:
(NewAccountModelPimpl::removeFromAccounts): The lock wait inside
`#ifdef CHK_FREEABLE_BEFORE_ERASE_ACCOUNT' was not updated when the
type of `accounts' was changed over the years from this:
  std::map<std::string, account::Info>
to this:
  std::map<QString, std::pair<account::Info, std::shared_ptr<Database>>>
Basically we need to get the `first' of the pair for `account::Info'.
So we now do that.

* src/libclient/avmodel.cpp:
* src/libclient/callbackshandler.cpp:
* src/libclient/contactmodel.cpp:
* src/libclient/conversationmodel.cpp:
* src/libclient/database.cpp:
* src/libclient/namedirectory.cpp:
* src/libclient/newaccountmodel.cpp:
* src/libclient/newcallmodel.cpp:
* src/libclient/newdevicemodel.cpp:
* src/libclient/peerdiscoverymodel.cpp:
* src/libclient/pluginmodel.cpp:
* src/libclient/smartinfohub.cpp:
* src/libclient/vcard.h:
* src/libclient/authority/storagehelper.cpp:  Replace Qt's `foreach' with
`Q_FOREACH' and its `emit' with `Q_EMIT', because in the client-qt
code base we have `-DQT_NO_KEYWORDS' to avoid conflicts with other
libraries we use.

GitLab: #734
Change-Id: I86e3b0fb30e554755023e7b858b6a0d132cd59ab