conferences: support multistream
In a conference, with this patch, the user is able to share
a media while keeping the camera. The mute button for the camera
only control the camera and we use requestMediaChange to add medias.
To stop sharing a media, the user will click on the dedicated button.
If the peer is not compatible with multistream, the behavior will
be the same as before

https://git.jami.net/savoirfairelinux/jami-project/-/issues/1429
Change-Id: If5e0634be6d0d06900ad82554b32a524fd4db36c