conversationmodel: avoid duplicated requests when receiving a call
Because the call and the conversation's requests are not related,
we can receive a call before the conversation's request. In this
case, a duplicate was created when accepting the call.
Moreover, accepting a call is now separated from accepting a contact,
like the android client (and accepting a conversation needs an explicit
action from the user).

Change-Id: I7c3eb34b7e121b63f1b8c99d084a5165054670d5