vcard: decode escaped chars
Since vcard message might arrive with encoded chars. To
avoid crashes we properly decode any message and check for
the values of interest within a fixed message format.

Note: Since https://github.com/pjsip/pjproject/pull/2933, the
part=/of= is not un-escaped anymore as the RFC allow escaped
character. So this was an old pjproject bug and was working with
luck.

Change-Id: If454cecab0c7ad6867680306687fb1452dc450cd