mainapplication: handle jami: URIs
Asking jami to launch another instance with ./jami-qt jami:xxx
will search for "xxx" inside the application and select the
conversation if found.
This allow to support webpage with <a href="jami:username">

For GNU/Linux, in the .desktop file, x-scheme-handler/jami is
added.
For packaging, we will need to determine if the scheme is handled
automatically. For now, to test, we need to add:

x-scheme-handler/jami=jami-qt.desktop;

in "~/.config/mimeapps.list" (and jami-qt.desktop) should
be in a valid path.

Change-Id: Ibbb6e8942f1873d81a57640bdf6f839885981be3
GitLab: #655