chat-view: fix data transfer image size reload loops
Changing the source size property of the QML Image component causes
a reload using the new source dimensions. The image loading process
was triggering reloading that was not recognized as a binding loop.

This commit also corrects the image sizing algorithm to prefer and
restrict height, which prevents images that are too tall from
taking up too much vertical space in the chat list view.

GitLab: #857
Change-Id: I049b1bb8ea4d23a753e7b54de884d9c1eafdf83c