packaging: Only rebuild source tarball when tarballs manifest changes.
This partially reverts commit 6e0a30024bd5d0f24f8c16960a56b02737bf0131
from the former jami-project repository ("build: Improvements to the
generation of the release tarballs."), which caused the source tarball
to always be rebuilt, even in the absence of changes.

This gets in the way when testing the various packaging targets, so
revert it.

* extras/packaging/gnu-linux/Makefile (release-tarball): Do not always
clear the tarballs.manifest file.  The purge-release-tarballs target
can be used for that.

Change-Id: I30ebda8dc258d0c95c6cf7e02a814627ac5b812b