tests: qml: initialize the preview engine
This was causing invalid connections during test setup. Not having a null preview engine will prevent dereferencing a null pointer once messaging tests are implemented, however, the webengine will not work correctly due to missing resources which may be corrected in a subsequent patch.

This commit also moves .clang-format to the root repo dir so it will apply to the test source code (the test source code should be within the src dir).

Gitlab: #899
Change-Id: I8cac1fb937a0c60a91219b11ed1e0b5f323b1816