misc: fix GTK theme application on Wayland by prioritizing XCB
This commit addresses an issue where the GTK theme was not being correctly applied when running our application on Wayland with the Wayland platform plugin. Although the root cause is not entirely clear, it seems related to a known Qt bug (QTBUG-99684). As a workaround, we're setting the `QT_QPA_PLATFORM` environment variable to use XCB with Wayland as a fallback. This ensures that the GTK theme is applied correctly when X11 is available, while still allowing a fallback to Wayland in its absence.

Possibly related issue: https://bugreports.qt.io/browse/QTBUG-99684

Gitlab: #1156
Change-Id: I71ef4e39f916d9871134fb0caded8b30bff578fb