msi-installer: add install scope choice
Permit using the advanced menu to select between per-user(default) and per-machine installation.

Note: the effects of migrating in between install scopes across upgrades is dissuaded by the WiX community, and has not been tested. At a minimum, over-installing the wrong scope will create another entry in the registry. The single-user and multi-user install instances can be managed separately in that case, but this may not be clear to the average user.

Gitlab: #1453
Change-Id: I228a2b9a48bce44d422ef612ffb57fd0342f8dc5