misc: only remove frameless system buttons when in a fullscreen call
We can't differentiate between F11 fullscreen and the window having geometry that matches its current display. This can happen if the user hides the start bar on Windows or otherwise manages to remove desktop decorations that restrict the maximization of the window.

This is a workaround until we can refactor or remove F11 support.

Change-Id: If8539f2b042eb14c9deb43aad76f7a7735243465