accountmodel: improve handling for account list reordering
This fixes a stale account list being used in AccountListModel by
synchronizing list invalidation to a newly introduced signal (accountsReordered). This prevents a data race that may occur, as
setting the account order is asynchronous over D-Bus, or when
queueing the signal handling for accountsChanged which is emitted
only once the reordering is saved.

Consequently, QAbstractItemModel mutations are now performed in AccountListModel instead of in the UI.

Gitlab: #1638 (Account list in popup is incorrect after selection)
Change-Id: I7ed6eeb45eb319f21e40554f3d023ad24e139a6f