LineEditContextMenu: fix for menu not respawnable
On the first right click on text in the chatbox(messagebartextarea), the
"cut, copy, paste" menu pops up. On subsequent right clicks, the menu
fails to spawn. The ancestor of the menu items, BaseContextMenu,
destroys them before terminating. Because the menuitems are only ever
created once, subsequent access attempts fail.

In this patch, the menuitems are created dynamically on a right click.

SFL: #1895
Change-Id: Ic01275a5645aebc974761b0ac4d4ee9969b9dd62