MessagesResearchView.qml: remove manual formatting
We should use the default formatting provided by Qt/QML components and
layout managers as much as possible. This is a follow-up to commit
82c876c0fa6aa8010da7af528219278e5b21d609 that tries to address the
overflow issue by making sure the components in the hierarchy have a
defined width, and using the TextEdit.wrapMode property to wrap the
text. To avoid very long messages to be displayed, we also limit the
height of the TextEdit to 4 lines with a fade-out effect for the text
that is too high instead of eliding it.

Some other small changes are added to make the component more robust:

- the current conversation ID change event is now used to update the
  prompt search;

- some extra spacing is added at the bottom of the view to avoid the
  last message to be glued to the bottom of the screen.

GitLab: #1827
Change-Id: Iec4e34c205f8537ab36c51dbf50c36efc4fc2744