settings: allow devs to modify default settings
Delete the default settings in the settings file so that only user
defined ones are stored locally.

Previously, we stored every value in the settings file. This created a problem where we could not modify a default value as we had no way of knowing if it was user defined or not.

This patch allow us to know that if a value was in the file, it's that it was different from the default.

GitLab: #2000
Change-Id: I7d5154b258c2ca97a1976dc006a70db0d5fe0b31