ci: migrate to Jenkinsfile for windows packaging
Repository currently uses a Jenkins script not part of the project's
source control, this commit aims to fix this.
Also, as there are now 2 jenkins Windows machines, the code signing is a
separate job triggered by the build job so it can be done on only one
machine even if the build can be run on both. For now, this code signing
is done through a Jenkins script, not a Jenkinsfile.
Changes:

- Add Jenkinsfile for windows packaging builds
- Remove code signing part of the pipeline to use it in another Jenkins
pipeline

Change-Id: I320a1a638c5bf46a5e06111a59d101f93a7f2ee1