spellcheck: windows and macos
Implement the hunspell spellchecker for Windows and MacOS. It also change the base implementation for Linux.
We now have a list of available dictionary and we fetch them from the LibreOffice repository.
It also refactor the precedent implementation into a spellchecker, a spellcheckhandler and a spellcheckdictionarymanager.

Additionaly, the context menu has been improved, we can now have up to 10 spelling recommandations for a word and the settings have been moved to the chat tab.

GitLab: #1997
Change-Id: Ibd0879f957f27f4c7c5720762ace553ca84e2bc3