spellcheck: windows and macos
Implement the hunspell spellchecker for Windows and MacOS. It also changes
the base implementation for Linux. The system dictionaries (if any) are
aggregated with those installed from the LibreOffice repository via Jami's
dictionary managemnet interface.

This commit implements a major refactoring of the spellcheck system to
improve UI responsiveness and user experience:

Core Changes:
- Used QAbstractListModel to represent the list of dictionaries
- Added new QML components:
  - DictionaryInstallView.qml
  - ManageDictionariesDialog.qml
  - SpellCheckLanguageComboBox.qml
- Updated property names for clarity
- Fixed a bug in the settings combo box custom component that caused
  out-of-range errors for filtered models

GitLab: #1997
Change-Id: Ibd0879f957f27f4c7c5720762ace553ca84e2bc3