spellcheck: windows and macos
Implement the hunspell spellchecker for Windows and MacOS. It also change the base implementation for Linux.
We now have a list of available dictionary and we fetch them from the LibreOffice repository.

This commit implements a major refactoring of the spellcheck system to
improve UI responsiveness and user experience:

Core Changes:
- Used QAbstractListModel to represent the list of dictionaries
- Added new QML components:
  - DictionaryInstallView.qml
  - ManageDictionariesDialog.qml
  - SpellCheckLanguageComboBox.qml
- Updated property names for clarity
- Fixed a bug in the settings combo box custom component that caused
  out-of-range errors for filtered models

GitLab: #1997
Change-Id: Ibd0879f957f27f4c7c5720762ace553ca84e2bc3