ci: migrate to Jenkinsfile for windows packaging
Repository currently uses a Jenkins script not part of the project's
source control, this commit aims to fix this.
Also, as there are now 2 jenkins Windows machines, the code signing is a
separate job triggered by the build job so it can be executed on the
machine that has the physical key. For now, this code signing
is done through a Jenkins script, not a Jenkinsfile.

Changes:
- Add Jenkinsfile for windows packaging builds
- Remove code signing part of the pipeline to use it in another Jenkins
pipeline

GitLab: #2047
Change-Id: I90d9051fba98edda2e5b94b1a0ba444142393f02