appsettings: refactor settings
The application settings are currently accessed in the app through the
Utilsadapter. This is an additionnal layer of code adding no value, so
this commit aims to refactor the settings management to use a
QmlPropertyMap stored in the appsettingsmanager. This lets the UI react
dynamically to changes made to the property map.
Changes :
- Migrate app settings access from UtilsAdapter to AppSettingsManager
- Update qml files to use the new QmlPropertyMap
- Move isRTL to AppSettingsManager

Change-Id: I920109ebb09fab8d74ef1217dd3c9ff82d147ac4