src/libclient/api/interaction.h: translate call duration
https://doc.qt.io/qt-6/qstring.html
The '%' can be followed by an 'L', in which case the sequence is
replaced with a localized representation of a. The conversion uses
the default locale, set by QLocale::setDefault(). If no default locale
was specified, the system locale is used. The 'L' flag is ignored if
base is not 10.
Please see the https://www.omniglot.com/language/numerals.htm
page for examples.

GitLab: #1979
Change-Id: I1f6bd06ef3c72cd395f45d8be40fa48f8df4022a