SelectScreen: align buttons to the right
This aligns the Cancel and Share Screen/Window buttons to the right instead of the left. The Share button is also now placed to the right of the Cancel button. This is also in preparation for the Audio sharing functionality that might require a Toggle Switch that will be placed on the bottom left for a natural flow of execution.

Change-Id: Iaaa0c39c1b58674c4c9d678caee89562ef67c53c