AppearanceSettings: add background blur slider and preview
This adds a slider in the appearance settings page for the user to blur the background image. The blur value is set in the account settings property map. The blur is done using FastBlur effect. GIF that demonstrates the feature:
https://mattermost.savoirfairelinux.com/api/v4/files/st3iezj3kpb5mrrowq57ncnmjy

This patch builds on top of https://review.jami.net/c/jami-client-qt/+/32067

Change-Id: I20334e52a7a19d98e5237f759570a118d7487cb9