MainOverlay: fix remote peer name still displayed in conference
A and B are in a call. From A's POV, B's name is displayed in the top
left. And for B, A's name is displayed in the top left. C joins the
call, making it a conference. The layout has changed, but for A and B
the names are still displayed in the top left and overlap with the
conference's UI. This patch fixes that by hiding this behavior when the
call is a conference.

Change-Id: Idc857d0b3e9ffaf99c829fd7cb0742f3592cca03